v0.3 release candidate
#14 - improving error messages when json is invalid.

Some progress on #13 and other errors - the way the threads were setup,
they continued without pausing if they hit an error. Now introduced
'pause' functionality on the thread, outside of the autoUpdate/runOnce
sphere that kicks in when an exception is hit, so that we don't spam
exceptions. Reloading/hitting 'update now' will reset that state and try
again.

Reworked handleException so that the apiKey is replaced with
YOUR_UNSANITISED_API_KEY_HERE - I had always said upload logs into
issues, but that is a _bad_ idea with those in. Logs should now be safe,
as are error messages.