#22 Update now hotkey
Not a complete solution, but this will allow users to do a request to
the webserver to cause an update on next tick, triggering the manual
update.

This can be any kind of request - PUSH, GET etc, which many automation
tools can do - such as streamdeck's default "Website" option which
performs a GET, which you can "Access in the background" to be
completely invisible and do exactly what you need.